Move slash command behavior to main.py, and remove extraneous srcfiles
